Move weapons specific code to weapons plugin
